Move bback some classes to admin_cohort and tools.py
